add PInvoke samples
Move the PInvoke sample to the new folder.
