+++ docs/csharp/advanced-topics/interop/snippets/WinSound/snippets5000.json
@@ -1,0 +1,3 @@
+{
+    "host": "visualstudio"
+}
