--- samples/snippets/csharp/VS_Snippets_VBCSharp/csProgGuideInterop/CS/WinSound.cs
@@ -1,133 +1,0 @@
-﻿//-----------------------------------------------------------------------------
-//<Snippet3>
-using System.Windows.Forms;
-using System.Runtime.InteropServices;
-
-namespace WinSound
-{
-    public partial class Form1 : Form
-    {
-        private TextBox textBox1;
-        private Button button1;
-
-        public Form1()  // Constructor.
-        {
-            InitializeComponent();
-        }
-
-        [DllImport("winmm.DLL", EntryPoint = "PlaySound", SetLastError = true, CharSet = CharSet.Unicode, ThrowOnUnmappableChar = true)]
-        private static extern bool PlaySound(string szSound, System.IntPtr hMod, PlaySoundFlags flags);
-
-        [System.Flags]
-        public enum PlaySoundFlags : int
-        {
-            SND_SYNC = 0x0000,
-            SND_ASYNC = 0x0001,
-            SND_NODEFAULT = 0x0002,
-            SND_LOOP = 0x0008,
-            SND_NOSTOP = 0x0010,
-            SND_NOWAIT = 0x00002000,
-            SND_FILENAME = 0x00020000,
-            SND_RESOURCE = 0x00040004
-        }
-
-        private void button1_Click(object sender, System.EventArgs e)
-        {
-            var dialog1 = new OpenFileDialog();
-
-            dialog1.Title = "Browse to find sound file to play";
-            dialog1.InitialDirectory = @"c:\";
-            //<Snippet5>
-            dialog1.Filter = "Wav Files (*.wav)|*.wav";
-            //</Snippet5>
-            dialog1.FilterIndex = 2;
-            dialog1.RestoreDirectory = true;
-
-            if (dialog1.ShowDialog() == DialogResult.OK)
-            {
-                textBox1.Text = dialog1.FileName;
-                PlaySound(dialog1.FileName, new System.IntPtr(), PlaySoundFlags.SND_SYNC);
-            }
-        }
-
-        private void Form1_Load(object sender, EventArgs e)
-        {
-            // Including this empty method in the sample because in the IDE,
-            // when users click on the form, generates code that looks for a default method
-            // with this name. We add it here to prevent confusion for those using the samples.
-        }
-    }
-}
-//</Snippet3>
-
-//-----------------------------------------------------------------------------
-// Form1.Designer.cs
-//-----------------------------------------------------------------------------
-namespace WinSound
-{
-    partial class Form1
-    {
-        /// <summary>
-        /// Required designer variable.
-        /// </summary>
-        private System.ComponentModel.IContainer components=null;
-
-        /// <summary>
-        /// Clean up any resources being used.
-        /// </summary>
-        /// <param name="disposing">true if managed resources should be disposed; otherwise, false.</param>
-        protected override void Dispose(bool disposing)
-        {
-            if (disposing&&(components!=null))
-            {
-                components.Dispose();
-            }
-            base.Dispose(disposing);
-        }
-
-        /// <summary>
-        /// Required method for Designer support - do not modify
-        /// the contents of this method with the code editor.
-        /// </summary>
-        private void InitializeComponent()
-        {
-            this.components=new System.ComponentModel.Container();
-            this.AutoScaleMode=System.Windows.Forms.AutoScaleMode.Font;
-            this.Text="WinSound";
-
-            //<Snippet4>
-            this.button1 = new System.Windows.Forms.Button();
-            this.textBox1 = new System.Windows.Forms.TextBox();
-            this.SuspendLayout();
-            //
-            // button1
-            //
-            this.button1.Location = new System.Drawing.Point(192, 40);
-            this.button1.Name = "button1";
-            this.button1.Size = new System.Drawing.Size(88, 24);
-            this.button1.TabIndex = 0;
-            this.button1.Text = "Browse";
-            this.button1.Click += new System.EventHandler(this.button1_Click);
-            //
-            // textBox1
-            //
-            this.textBox1.Location = new System.Drawing.Point(8, 40);
-            this.textBox1.Name = "textBox1";
-            this.textBox1.Size = new System.Drawing.Size(168, 20);
-            this.textBox1.TabIndex = 1;
-            this.textBox1.Text = "FIle path";
-            //
-            // Form1
-            //
-            this.AutoScaleDimensions = new System.Drawing.SizeF(5, 13);
-            this.ClientSize = new System.Drawing.Size(292, 266);
-            this.Controls.Add(this.textBox1);
-            this.Controls.Add(this.button1);
-            this.Name = "Form1";
-            this.Text = "Platform Invoke WinSound C#";
-            this.ResumeLayout(false);
-            this.PerformLayout();
-            //</Snippet4>
-        }
-    }
-}
